common.h: Add ReGBA_GetBackupFilename and ReGBA_GetSavedStateFilename.
DSTwo, GCW Zero: Create implementations of the above functions. GCW Zero: Use .sNN (.s00 through .s99) for saved state file names. Delete get_savestate_filename and get_savestate_filename_noshot, replacing them with the new functions.

memory.c: Delete references to DEFAULT_SAVE_DIR, a DSTwo port variable. Delete gamepak_filename. Replace it with CurrentGamePath, uninitialised, and IsGameLoaded, initialised to false. This makes it so the full path of the game is kept around, and initialising the first character of gamepak_filename is not needed anymore. Make load_backup, save_backup, load_state and save_state not require a path to be passed, only requiring the slot number now and using the functions above to resolve the file names per platform.

+++ source/zip.h
@@ -26,7 +26,7 @@
 
 #define ZIP_TMP "ZIPTMP.GBA" /* Acceptable on all platforms */
 
-s32 load_file_zip(char *filename);
+ssize_t load_file_zip(char *filename);
 
 #endif
